Fixing the issue that an invalid public key could be submitted
A bad actor could submit a completely bogus key when uploading a file, since it's encrypted and there isn't a way to validate encrypted data. We fix this by making the public key no longer encrypted at all, since the hosting node would have to know the public key anyway to check the reputation for the PoW challenge, and ensuring that a file can't be tied back to a user unless exceptional circumstances isn't one of the points of this project.